rename files and methods to remove repetitive wording
