+++ pkg/api/2018-09-30-preview/converterto_test.go
@@ -1,4 +1,4 @@
-package v20190430
+package v20180930preview
 
 import (
 	"reflect"
@@ -9,22 +9,19 @@ import (
 	"github.com/openshift/openshift-azure/pkg/api"
 )
 
-// api.GetInternalMockCluster and v20190430ManagedCluster are defined in
-// converterfromv20190430_test.go.
-
-func TestConvertTov20190430(t *testing.T) {
+func TestConvertTo(t *testing.T) {
 	tests := []struct {
 		cs *api.OpenShiftManagedCluster
 		oc *OpenShiftManagedCluster
 	}{
 		{
 			cs: api.GetInternalMockCluster(),
-			oc: v20190430ManagedCluster(),
+			oc: managedCluster(),
 		},
 	}
 
 	for _, test := range tests {
-		oc := ConvertTov20190430(test.cs)
+		oc := ConvertTo(test.cs)
 		if !reflect.DeepEqual(oc, test.oc) {
 			t.Errorf("unexpected result:\n%#v\nexpected:\n%#v", spew.Sprint(oc), spew.Sprint(test.oc))
 		}
