+++ pkg/api/2019-04-30/converterto_test.go
@@ -1,4 +1,4 @@
-package v20180930preview
+package v20190430
 
 import (
 	"reflect"
@@ -9,22 +9,19 @@ import (
 	"github.com/openshift/openshift-azure/pkg/api"
 )
 
-// api.GetInternalMockCluster and v20180930previewManagedCluster are defined in
-// converterfromv20180930preview_test.go.
-
-func TestConvertToV20180930preview(t *testing.T) {
+func TestConvertTo(t *testing.T) {
 	tests := []struct {
 		cs *api.OpenShiftManagedCluster
 		oc *OpenShiftManagedCluster
 	}{
 		{
 			cs: api.GetInternalMockCluster(),
-			oc: v20180930previewManagedCluster(),
+			oc: managedCluster(),
 		},
 	}
 
 	for _, test := range tests {
-		oc := ConvertToV20180930preview(test.cs)
+		oc := ConvertTo(test.cs)
 		if !reflect.DeepEqual(oc, test.oc) {
 			t.Errorf("unexpected result:\n%#v\nexpected:\n%#v", spew.Sprint(oc), spew.Sprint(test.oc))
 		}
