+++ pkg/api/admin/converterfrom_test.go
@@ -1,6 +1,7 @@
-package v20180930preview
+package admin
 
 import (
+	"crypto/x509"
 	"errors"
 	"reflect"
 	"testing"
@@ -9,10 +10,11 @@ import (
 	"github.com/go-test/deep"
 
 	"github.com/openshift/openshift-azure/pkg/api"
+	"github.com/openshift/openshift-azure/pkg/util/tls"
 	"github.com/openshift/openshift-azure/test/util/populate"
 )
 
-func v20180930previewManagedCluster() *OpenShiftManagedCluster {
+func managedCluster() *OpenShiftManagedCluster {
 	// use populate.Walk to generate a fully populated
 	// OpenShiftManagedCluster
 
@@ -30,7 +32,26 @@ func v20180930previewManagedCluster() *OpenShiftManagedCluster {
 	return &omc
 }
 
-func TestConvertFromV20180930preview(t *testing.T) {
+func internalManagedCluster() *api.OpenShiftManagedCluster {
+	cs := api.GetInternalMockCluster()
+
+	prepare := func(v reflect.Value) {
+		switch v.Interface().(type) {
+		case []api.IdentityProvider:
+			// set the Provider to AADIdentityProvider
+			v.Set(reflect.ValueOf([]api.IdentityProvider{{Provider: &api.AADIdentityProvider{Kind: "AADIdentityProvider"}}}))
+		}
+	}
+	populate.Walk(&cs, prepare)
+	return cs
+}
+
+func TestConvertFrom(t *testing.T) {
+	_, dummyCA, err := tls.NewCA("dummy")
+	if err != nil {
+		t.Fatal(err)
+	}
+
 	tests := []struct {
 		name           string
 		input          *OpenShiftManagedCluster
@@ -38,10 +59,6 @@ func TestConvertFromV20180930preview(t *testing.T) {
 		expectedChange func(*api.OpenShiftManagedCluster)
 		err            error
 	}{
-		{
-			name:  "create",
-			input: v20180930previewManagedCluster(),
-		},
 		{
 			name: "router profile update",
 			input: &OpenShiftManagedCluster{
@@ -54,7 +71,7 @@ func TestConvertFromV20180930preview(t *testing.T) {
 					},
 				},
 			},
-			base: api.GetInternalMockCluster(),
+			base: internalManagedCluster(),
 			expectedChange: func(expectedCs *api.OpenShiftManagedCluster) {
 				expectedCs.Properties.RouterProfiles[0].PublicSubdomain = "NewPublicSubdomain"
 			},
@@ -70,7 +87,7 @@ func TestConvertFromV20180930preview(t *testing.T) {
 					},
 				},
 			},
-			base: api.GetInternalMockCluster(),
+			base: internalManagedCluster(),
 			err:  errors.New("invalid router profile - name is missing"),
 		},
 		{
@@ -89,7 +106,7 @@ func TestConvertFromV20180930preview(t *testing.T) {
 					},
 				},
 			},
-			base: api.GetInternalMockCluster(),
+			base: internalManagedCluster(),
 			expectedChange: func(expectedCs *api.OpenShiftManagedCluster) {
 				expectedCs.Properties.AgentPoolProfiles = append(expectedCs.Properties.AgentPoolProfiles,
 					api.AgentPoolProfile{
@@ -117,7 +134,7 @@ func TestConvertFromV20180930preview(t *testing.T) {
 					},
 				},
 			},
-			base: api.GetInternalMockCluster(),
+			base: internalManagedCluster(),
 			err:  errors.New("invalid agent pool profile - name is missing"),
 		},
 		{
@@ -129,14 +146,14 @@ func TestConvertFromV20180930preview(t *testing.T) {
 							{
 								Name: to.StringPtr("Properties.AuthProfile.IdentityProviders[0].Name"),
 								Provider: &AADIdentityProvider{
-									Secret: to.StringPtr("NewSecret"),
+									ClientID: to.StringPtr("NewClientID"),
 								},
 							},
 						},
 					},
 				},
 			},
-			base: api.GetInternalMockCluster(),
+			base: internalManagedCluster(),
 			expectedChange: func(expectedCs *api.OpenShiftManagedCluster) {
 				expectedCs.Properties.AuthProfile = api.AuthProfile{
 					IdentityProviders: []api.IdentityProvider{
@@ -144,9 +161,43 @@ func TestConvertFromV20180930preview(t *testing.T) {
 							Name: "Properties.AuthProfile.IdentityProviders[0].Name",
 							Provider: &api.AADIdentityProvider{
 								Kind:     "AADIdentityProvider",
-								ClientID: "Properties.AuthProfile.IdentityProviders[0].Provider.ClientID",
-								Secret:   "NewSecret",
+								ClientID: "NewClientID",
+								Secret:   "Properties.AuthProfile.IdentityProviders[0].Provider.Secret",
 								TenantID: "Properties.AuthProfile.IdentityProviders[0].Provider.TenantID",
+							},
+						},
+					},
+				}
+			},
+		},
+		{
+			name: "auth profile update aad groups",
+			input: &OpenShiftManagedCluster{
+				Properties: &Properties{
+					AuthProfile: &AuthProfile{
+						IdentityProviders: []IdentityProvider{
+							{
+								Name: to.StringPtr("Properties.AuthProfile.IdentityProviders[0].Name"),
+								Provider: &AADIdentityProvider{
+									CustomerAdminGroupID: to.StringPtr("admin"),
+								},
+							},
+						},
+					},
+				},
+			},
+			base: internalManagedCluster(),
+			expectedChange: func(expectedCs *api.OpenShiftManagedCluster) {
+				expectedCs.Properties.AuthProfile = api.AuthProfile{
+					IdentityProviders: []api.IdentityProvider{
+						{
+							Name: "Properties.AuthProfile.IdentityProviders[0].Name",
+							Provider: &api.AADIdentityProvider{
+								Kind:                 "AADIdentityProvider",
+								ClientID:             "Properties.AuthProfile.IdentityProviders[0].Provider.ClientID",
+								Secret:               "Properties.AuthProfile.IdentityProviders[0].Provider.Secret",
+								TenantID:             "Properties.AuthProfile.IdentityProviders[0].Provider.TenantID",
+								CustomerAdminGroupID: to.StringPtr("admin"),
 							},
 						},
 					},
@@ -169,7 +220,7 @@ func TestConvertFromV20180930preview(t *testing.T) {
 					},
 				},
 			},
-			base: api.GetInternalMockCluster(),
+			base: internalManagedCluster(),
 			err:  errors.New("cannot update the kind of the identity provider"),
 		},
 		{
@@ -187,27 +238,60 @@ func TestConvertFromV20180930preview(t *testing.T) {
 					},
 				},
 			},
-			base: api.GetInternalMockCluster(),
+			base: internalManagedCluster(),
 			err:  errors.New("invalid identity provider - name is missing"),
 		},
 		{
-			name: "nil ResourcePurchasPlan update",
+			name: "image offer update",
 			input: &OpenShiftManagedCluster{
-				Plan: nil,
+				Config: &Config{
+					ImageOffer: to.StringPtr("NewOffer"),
+				},
 			},
-			base: api.GetInternalMockCluster(),
+			base: internalManagedCluster(),
 			expectedChange: func(expectedCs *api.OpenShiftManagedCluster) {
+				expectedCs.Config.ImageOffer = "NewOffer"
+			},
+		},
+		{
+			name: "certificate update",
+			input: &OpenShiftManagedCluster{
+				Config: &Config{
+					Certificates: &CertificateConfig{
+						OpenShiftConsole: &CertificateChain{
+							Certs: []*x509.Certificate{dummyCA},
+						},
+					},
+				},
+			},
+			base: internalManagedCluster(),
+			expectedChange: func(expectedCs *api.OpenShiftManagedCluster) {
+				expectedCs.Config.Certificates.OpenShiftConsole.Certs = []*x509.Certificate{dummyCA}
+			},
+		},
+		{
+			name: "loglevel update",
+			input: &OpenShiftManagedCluster{
+				Config: &Config{
+					ComponentLogLevel: &ComponentLogLevel{
+						Node: to.IntPtr(2),
+					},
+				},
+			},
+			base: internalManagedCluster(),
+			expectedChange: func(expectedCs *api.OpenShiftManagedCluster) {
+				expectedCs.Config.ComponentLogLevel.Node = to.IntPtr(2)
 			},
 		},
 	}
 
 	for _, test := range tests {
-		expected := api.GetInternalMockCluster()
+		expected := internalManagedCluster()
 		if test.expectedChange != nil {
 			test.expectedChange(expected)
 		}
 
-		output, err := ConvertFromV20180930preview(test.input, test.base)
+		output, err := ConvertFrom(test.input, test.base)
 		if !reflect.DeepEqual(err, test.err) {
 			t.Errorf("%s: expected error: %v, got error: %v", test.name, test.err, err)
 		}
@@ -219,13 +303,13 @@ func TestConvertFromV20180930preview(t *testing.T) {
 	}
 }
 
-func TestRoundTripV20180930preview(t *testing.T) {
-	start := v20180930previewManagedCluster()
-	internal, err := ConvertFromV20180930preview(start, nil)
+func TestRoundTrip(t *testing.T) {
+	start := managedCluster()
+	internal, err := ConvertFrom(start, nil)
 	if err != nil {
 		t.Error(err)
 	}
-	end := ConvertToV20180930preview(internal)
+	end := ConvertTo(internal)
 	if !reflect.DeepEqual(start, end) {
 		t.Errorf("unexpected diff %s", deep.Equal(start, end))
 	}
