+++ pkg/api/plugin/converterfrom.go
@@ -6,9 +6,11 @@ import (
 	"github.com/openshift/openshift-azure/pkg/api"
 )
 
-// ConvertFromPlugin converts external plugin API config type into
-// internal API Config type
-func ConvertFromPlugin(in *Config, old *api.Config, version string) (*api.Config, error) {
+// ConvertFrom converts from a
+// plugin.OpenShiftManagedCluster.Config to an internal.OpenShiftManagedCluster.Config
+// If old is non-nil, it is going to be used as the base for the internal
+// output where the external request is merged on top of.
+func ConvertFrom(in *Config, old *api.Config, version string) (*api.Config, error) {
 	if _, found := in.Versions[version]; !found {
 		return nil, fmt.Errorf("version %q not found", version)
 	}
